requirements.txt
numpy==1.23.1
opencv-python==4.5.5.64
pillow==9.2.0